// frontend/src/app/model/user.model.ts
export class User{
    username: string;
    mail: string;
    password: string;
    pwchg: number;
    name: string;
    lastname: string;
    address: string;
    phone: string;
    title: string;
    office: string;
    status: string;
    picture: string;
    bio: string;
    website: string;
    type: string;
    predmeti:Array<String>;
}